// typings.d.ts
export interface Post {
    _id: string;
    title: string;
    _createdAt: string;
    author: {
        name: string;
        image: string;
    };
    description: string;
    mainImage: {
        asset: {
            url: string;
        };
    };
    slug: {
        current: string;
    };
    body: [object];
}